requirements.txt
PyQt5==5.15.11
scapy==2.5.0
PyYAML==6.0.2